requirements.txt
django-heroku==0.3.1
gunicorn==20.0.4
whitenoise==5.2.0
